I may have made a change others don't like, so speak up if you feel like having a holy war over string lengths and null terminators... :)
A major bottleneck in a compiler is string compares.  I noticed that
String::strcmp is, well, about as inefficient as you could possibly
make it, but that String::zstrcmp is much better.  (The latter still
has lots of unnecessary array-bounds checks, but it doesn't have
indirect function calls and two passes.)  The difference is zstrcmp
doesn't stop at null characters.

But hey, nothing we compare in the compiler has null characters, so I
should be able to replace all the strcmp calls with zstrcmp, right?
<Insert demonic Dan-wasted-an-evening-making-this-true laughter.>

Prior to this check-in, the constant Cyclone string (string)"abc"
would become:
      (4,|)
         |
         V
         (a, b, c, \0)

But if the lexing engine decided the next token was abc, then Lexing::lexeme
conspire with runtime-function new_string to return:
      (3,|)
         |
         V
         (a, b, c, \0)

Alas, these two strings are equal according to strcmp, but not
according to zstrcmp.  Makes it awfully hard to recognize keywords,
doesn't it?

The bottom line is that I made two changes and fixed any affected code:
A. The size of constant strings no longer includes the null terminator.
B. new_string no longer puts in a null terminator.

So our examples would be
      (3,|)
         |
         V
         (a, b, c, \0)
and
      (3,|)
         |
         V
         (a, b, c)
which are equal according to strcmp and zstrcmp.  We could choose
to eliminate the \0 on constant strings too -- all we have to do
is translate "abc" to {'a', 'b', 'c'}.  I'm extremely tempted.
The other reasonable option is to always put in a "ghost \0".
Personally, I think programmers can put in their own.

Does it matter?  I changed a lot of strcmp to zstrcmp and the time for
bin/cyclone.exe to create src/cyclone.exe dropped from 79s to 67s on my
machine.

The real war story: Notice that change (A) happens at compile time
whereas change (B) happens at run time.  So in a bootstrapping system,
your head starts spinning.  You have to make change (A) one
iteration before you change the code that tests the size of constant
strings.  I couldn't keep things straight, so I ended up making change
(A) directly on the C code in bin/genfiles at "the same time" I
changed the Cyclone code.  Confusing, eh?

For future reference, here's a list of places to look in the code if
you go screwing with string representation:
  cyc_runtime: new_string, xprintf, string_to_Cstring,  Cstring_to_string
  lexing: aux_buffer, lex_refill, lexeme
  string: <duh>
  lex: token_string, string_buffer
  tcexp: String_c
  toc: TagFixedArray, generating xenum tags

Finally, I changed the library in a couple places to use zstrcmp
instead of strcmp. In pp, the hash table uses zstrcmp.  I believe this
change is safe even if "Doc"s have \0 characters because we'll just
lose some sharing.  In fact, I think you could argue this is a bug
fix.  In id, I'd have to claim the module just now assumes that
strings don't have \0 characters.