A mid-stream check-in -- control-flow checking (previously known as, "the subsequent pass") is partially implemented.  More specifically, crawling over statements is implemented, but crawling over expressions is still an "accept everything" stub.  The pass has already required some changes to the abstract syntax.  Notice that stmt now has an xenum so passes can annotate however they want.  This "open architecture" feels right, but enums are a lot more efficient.
We do now check that a goto does not jump into the scope of local
declarations or an exception handler.  That means once definite
assignment is properly implemented (i.e. I do the exps), we can
stop warning that goto may be unsafe.

All this new stuff is in the new file control_flow.cyc.  I think I
remembered to add all the right stuff to the repository.

Mostly I'm doing this check-in for other bug fixes and improvements
that I want to start exploiting.  None of them are very interesting.
I did add some more "fast paths" to the Set module.