Two fairly independent bug-fixes to how we treat type declarations.
First, I think I introduced a bug some weeks back wherein we weren't
actually checking struct/enum re-declarations against previous declarations.
You see, there's this hack where we put in an abstract version for
checking recursive instances.  You better not do that if there's already
a concrete version.

Second, we were doing a lot of relative-qvar lookups where they should have
been absolute.  The problem is that making a qvar absolute is not idempotent,
so you have to remember where you've done it.  Trevor is right that a much
more robust solution is to record in each qvar whether it's absolute or
relative.

But for now, I wanted pointers from struct/enum/xenum types to the
corresponding structdecl/enumdecl/xenumdecl.  In these cases, the compiler
now has the invariant that when the field is set, the name in the
struct/enum/xenum type is absolute.  check_valid_type sets these fields.
It does so with a level of indirection so that if an abstract type
becomes concrete, the concrete version will be available through the
StructType.  Therefore, we can just follow these pointers in, for example,
type-checking StructMember, rather than doing a dictionary lookup.
(That lookup had been relative when it should have been absolute.)

A little region stuff is in the code base now.  Please ignore it. :-)