Ok, I hope I'm supposed to be committing all of these genfiles.  I've never updated the libraries before...it bootstraps and passes the tests...
Here are a bunch of changes, essentially a big addition
to the Array library.

There are now array versions of most of the functions
in the List library (map, app, iter, folds, etc --
everything in lib/array.h except qsort is new, have a
look).  The comments are often lifted from the List
library (I adopted the SML/NJ notation for arrays as
[|...|] for lack of any better idea; I'm certainly not
attached to it).

In addition, Array::msort (same prototype as qsort) is
mergesort, which is stable.  I tried to code it
pretty efficiently (at least for large cases -- there's
some overhead I didn't try to minimize), so although I
doubt it can compete with mergesort it should be usable
if anyone has use for a stable sort.

Bugfix in array bounds check in Array::qsort.

Minor changes to List library.  Fixed an incorrect
comment and added List::exists_c which was missing.

Modified tests/cyctest.cyc a bit.  There's now pr_array
and array_cmp, analogous to pr_list and list_cmp
(should the comparison functions be put in the
libraries?).  There are a few more test cases for
arrays, but nowhere near enough.  Anybody really fond
of writing test cases?