Added some tests for arrays of something more complicated than int.  This confirms an confusion I'm having.
Say we have, in Cyclone,

int f<`a>(`a x[?]) {
  return x.size;
}

int g<`a>(`a x[?]) {
  return sizeof(x);
}

.  Greg expected that they would return different results
(in particular, f returns number of elements and g returns
size of x, in byes).  However, in practice, it seems that
they are both returning the number of elements.  (See near
the end of the array section of tests/cyctest.cyc for
examples.)

(In fact, calling sizeof on an int gets me "size requires
array type, not int".)

Bug, or Greg being confused?  And if this is the desired
behavior, how should one decide which idiom to use?