I got tired of our parser-generated files looking quite so ugly. So now I generate a destructor for each YYSTYPE variant and translate $n to the appropriate function call.  For example,
_ yyget_TypeList_tok(xenum YYSTYPE x) {switch(x) {case TypeList_tok(y):
return y; default: throw Core::Failure("TypeList_tok");}}

Note that type inference saves our butt (the return type is inferred)
just as it used to with the inlined dereference.

As a result parse_tab.c is about 30% smaller and the shrinkage comes
in the big yyparse function.

There is one change is what compiles as a result: Suppose you have
a YYSTYPE variant that carries 0 or more than 1 value.  It used to be
you could use this variant in a %type declaration but you could not use
$n such that the type was this variant.  You now cannot use the variant
in a %type declaration.

I changed the one parse.y variant that was affected by this dark corner.

I have to admit, though, the files still look pretty ugly. :-)