I wasn't checking that regions were accessible when accessed. So this code will now fail to type-check, as it should
int main() {
  int *_ x = null;
  if(1) {
    int y = 17;
    x = &y;
    *x = 75; // safe
  }
  if(x != null)
    *x = 99; // not safe!
  return 0;
}

Basically, a PointerType has a region field and you should call
Tcenv::check_rgn_accessible with it and the tenv if the pointer will
be followed (tcDeref, tcStructArrow, tcSubscript).  Fenv now keeps
track of all the blocks that are available.  May be buggy.

Forgot that the tcComprehension variable must be put in a new block.
I like comprehensions, but it's a pain to remember that this is the one
expression form with a binding occurrence.