Ok, this should make compiling a little cleaner (added a bunch of explicit casts from @ to * and $(t) to t to get it to shut up).  I think Dan is going to eliminate the warning about * patterns, so those are still there for now, but there shouldn't be any others.
There is one last warning about implicit casts in cyclex,
but it's not obvious to me how to fix it.  (Also in cyclex,
there were conflicting header definitions so I renamed them
so it wouldn't complain.)

Oh, and I reformatted a couple lines that were over 80
characters, while I was wandering around, and fixed a typo
in the documentation.