Ok, changed output format.  Two new flags, -q (quiet) and -v (verbose).  With -q, output nothing unless there's an error of some sort (make still outputs two lines/file).  With -v, output is old-style (one line per stage).  I didn't see any easy way to check what order arguments were in, so -v always overrides -q regardless of their order.
The new default, neither -q nor -v, prints

    Compiling file: foo.....

with one dot for each stage.

I didn't change any of the Makefiles, so now all of the
output is in the terser format (above).  If we want to
change the default behavior and put a flag in the Makefiles,
that's easy enough.

Do we return a useful errorcode, so users can check for
successful compile even with -q?