Added support for C++-style extern "C" declarations.  For instance:
  extern "C" double cos(double);

and

  extern "C" {
    double cos(double);
    double sin(double);
  }

At the Cyclone level, extern "C" is orthogonal to namespaces.
So, for instance, you can write:

  namespace CMath {
  extern "C" {
   double cos(double);
   double sin(double);
  }}

and refer to the C versions as CMath::cos and CMath::sin.  This
is consistent with the approach taken by C++.  It's very useful
for writing wrappers.  For instance, we could now write:

  namespace Math {
    double cos(double x) {
     // do some error checking
     CMath::cos(x);
     // do some more error checking
    }
  }

During translation, I drop the namespace on extern "C" variables.
This, the CMath declarations above will translate to:

  extern double cos(double);
  extern double sin(double);

This means that there is a potential for name-clashes that won't
be caught by the front-end.  However, I plan to (sort of) fix
this by wrapping an implicit Cyc namespace around all top-level
definitions.  We should really check that there's not already
something at the top-level with the same name.  I'll try to fix
all of this up in the next check in.