I added some new libraries, namely stdio, math, and assert. All of the file functionality that was in precore and core has been moved into stdio.  Thus, you'll need to #include "stdio.h" for just about any file now (hence the massive changes.)  In addition, cyc_stderr, cyc_stdout, cyc_stdin become Cyc::Stdio::stderr, etc.  So, using Stdio means they look like stderr, stdout, stdin, etc.
I wrote stdio.cyc as a bunch of wrappers to the C versions
of the underlying functions.  It's a good example of how to
use namespaces and extern "C" to write these wrappers.  I
put in almost all of the functions that are provided by C's
stdio, but left a few out that I didn't understand how to
make safe (e.g., gets).  Also, I left many of the old Cyclone
versions of file functions around (e.g., file_open).  The
Cyclone versions (which start with file_ instead of f) raise
exceptions instead of returning an error code.  We may want
to write more Cyclone versions for some of the functions or
do more error checking in the wrappers we have.

Math was a breeze to port -- just made it all extern "C".
However, we get a gcc warning that the struct __exception
declared in there (for reporting math exceptions) has a
"useless keyword or type name".  For the life of me, I can't
figure out why.  This shows up when compiling string.cyc
which is the only file that currently uses math.h (for the
pow function.)

I ported the assert library from Plaugher's book.  The only
difference is that I made assert return an int instead of
void -- this turned out to be necessary since we can't cast
any value (e.g., 0) to void.

Neither the math nor the stdio libaries have been tested
extensively -- they should be.  I have tested the assert
library.

We might want to create an include/ directory which has
dummy include files that suck in the library versions of
these header files.