Whew!  I changed the representation of t[?] arrays -- it's now something like:
  struct { t *curr; t *base; t *last_plus_one; }

which means that (a) it's unboxed, and (b) we can support
pointer arithmetic on these values a la C.  I haven't yet
added in the arithmetic but this should be fairly easy.

Changing the representation turned out to be hairy mostly
because string = char[?] is no longer a boxed type and we
used string in a lot of lists, dictionaries, hash tables,
etc.  So, I had to manually box things in those situations.
(You'll see some macros in cyc_include and precore that
helped with this a bit -- also some stuff from the old
string representation that I need to get rid of now.)

One bug that took me forever to track down was due to the
lack of link-checking.  So, I would suggest that we bump
up the priority for this.  In the meantime, I would strongly
suggest never manually declaring extern values, but rather
put them in a .h file and make sure the .h file is #included
in the appropriate implementation as well as the clients.
(parser.y declares some stuff from lex.cyl -- that's where
I ran into trouble.)