An enormous number of changes which I'll only highlight here:
1. completely re-worked the support for regions and region inference.
The old system (based on kinds) didn't really work right with
existentials and wouldn't work for closures at all.  This one
is based more upon Tofte & Talpin and the kind system is greatly
simplified as a result.

Each function type has an effect, which is conceptually a set of
regions that the function might access.  Effect sets can be abstracted
which is necessary to support higher-order functions.  You put
an effect on a function by putting a ";" (semi-colon) after
the arguments.  Effects can look like this basically:

    eff ::=   {t1,...,tn}  // ti are regions
            | eff + eff    // union of effects
            | `e           // an effect variable

Unlike Tofte & Talpin, we do not infer the effect for a function
by looking at its implementation.  Rather, we assume a "default"
effect which is a type-directed translation applied to the type
of the function.  This gives us separate compilation and mixes
nicely with explicit typing.  That is, if you don't like the default
effect, you can write your own effects.  Roughly speaking, the
default effect includes any free region you mention in the type of
the function.  It also includes any free effect variables for any
function types within the function.

The default effect inserts an effect variable for every function
type within a type.  This is necessary to get the unification-based
inference for effect instantiation to work well.  Rather than
have the user explicitly abstract these generated type variables,
we've moved to a setting where we provide ML-style automatic
prenex quantification of type variables.  As a result, what
you actually have to write for a given function's type is smaller,
as long as you're happy with the defaults.  For instance, the
type of the map function can be written simply like this:

  list_t<`b> map(`b f(`a), list_t<`a> x)

The "real" type that gets inferred as the default looks more like
this:

  list_t<`b> map<`a::B,`b::B,`e1:E,`e2::E>(`b f(`a;`e1), list_t<`a> x;`e1+`e2)

You can see that we don't want users having to write this crap.
But sometimes, it's necessary, particularly when you want to use
existentials or first-class polymorphism, or when the default
effect just is too conservative.

So, the upshot is, don't write explicit type abstractions on
functions unless you know what you're doing with the effect
system.  There's little need to, anyway.

2. In addition to figuring out default effects, we also infer the
kinds of type variables---B is boxed type, M is mem type, R is
region, and E is effect.  For abstract types (e.g., set_t<`a>),
you need to put kinds in explicitly unless they are B which is
the default.  If a type variable doesn't get used, then you'll
get a warning about it being underconstrained.

3. Structs, typedefs, and enums can be parameterized by any kind
now, not just boxed types.  This is necessary to support the
regions and the effects well.

4. Mem types and abstract structs can't occur everywhere --- they
have to occur under a pointer.  This functionality isn't completely
done right yet.

5. I've updated almost all of the libraries with more region and
effect-aware interfaces (mostly to stress-test the code.)  At the
same time, I tried to make the naming conventions more uniform.
For instance, it's now list_t<`a> instead of list<`a> and it's
&List instead of &Cons.  Similarly, it's dict_t<`a,`b> and
set_t<`a> and so forth.

In actuality, all of the libraries have a more general region-ized
version of the ADT.  For instance, for lists, we have:

  struct List<`a,`r> {`a hd; struct List<`a,`r> * `r tl;};

  typedef struct List<`a,`r> * `r glist_t<`a,`r>;

  typedef glist_t<`a,`H> list_t<`a>;

So, a List is parameterized by the type and the region for the tail.
A glist_t (general list) is a pointer to a List with the same region
as the List's tail.  (You could generalize this one more by allowing a
different region for the first pointer, but that wouldn't be too
useful.)  A list_t is a heap-allocated list.  Note that `H is the heap
region type (and not really considered a type variable at all.)

Most of the functions in the list library take glist_t's as
arguments.  Note that you can always pass a list_t to these
functions, so client code didn't really have to change.  Though
most of the arguments are glists, most of the results are
list_t's simply because I haven't added a way to allocate
into a region yet.  So, for instance, map really has the interface:

  list_t<`b>  map(`b f(`a), glist_t<`a,`r>);

Note that when we do add let-region and region handles, the
interface would change anyway to something like this:

  glist_t<`b,`r2>  gmap(`b f(`a), handle(`r2), glist_t<`a,`r1>);

That is, you'll have to pass in a region handle to allocate
the result.  So, I expect that we'll actually have a "gmap"
function which takes region handles, and a "map" function
which doesn't.  That way, clients that just want to heap
allocate don't have to keep passing the heap handle around.

There are probably many bugs still lurking in this code, and
I haven't tested many negative examples.  But the compiler
bootstraps and I wanted to get all of this code in by Spring
break.