Made translation of reference patterns to C less buggy.
Basically, it was broken if a refernce pattern appeared as a field to
a struct or tuple pattern that was not directly underneath a pointer
or at the top-level of the pattern.

You might want to read that last sentence again. :-)

Anyway, it now gets it correct, I think, and nothing broke.

BTW, you'll get a type error for:
switch(3) {
 case *x: ...
}
If this did type-check, you'd end up taking the address of a
compiler-generated temporary, which is probably not what you'd expect
in the case
switch(y) {
case *x: // x != &y
}