The old control flow files are no longer used or built by the compiler.  I have left them in the repository for now.
My last check-in left debugging (-g) on in the src and lib directories
which makes builds take about 20% lonnger.  Sorry about that.

The new control flow now does the check that you don't goto into a
scope or exception handler.

Fixed an unsoundness in taking the join of flowinfos.  Now as we take
the join we collect a set of places that escape.  After joining the
two flows, we go back through and deal with the consequences of
everything that escaped.  (Before we tried to adjust things "on the
fly" and that's not right.)

Fixed a bug in computing the initialization state of *e.  If e is
ThisIL, then *e is clearly NoneIL, but we used to say ThisIL.

Other tweaks are minor.