Added support for some GCC-style attributes, added an errno library, and added a new Structfield struct for struct fields.
The attributes have not been tested extensively and I'm sure there
are bugs and differences with GCC.  In particular, GCC seems to
allow attributes to go just about anywhere in a declaration.  We don't
allow them in as many places in order to avoid parser conflicts.  But
I added them just about anywhere I possibly could otherwise.

You use __attribute__((<attribute-list>)) on a declaration to
add the attributes.  Right now, we only support attributes on:
(a) function declarations, (b) variable declarations, (c) structs,
(d) struct fields as detailed below.

I've only checked that the "dllimport" attribute works for
functions and variables as I needed this for the errno library.

The attributes that we supposedly support are as follows:

For function types:
  stdcall -- x86 only, callee pops arguments (used by system calls)
  cdecl   -- x86 only, caller pops arguments
  noreturn -- function does not return (e.g., exit).  The type-checker
              doesn't take this into account but could in the future
              if it's willing to enforce it.
  const -- function doesn't mutate anything but local variables.  For
           optimization purposes.  The type-checker neither uses nor
           enforces this.
  regparm(i) -- x86 only, the first i scalar arguments are passed in registers
                (i in the range [0..3])

For variable declarations, structs, and struct fields:
  aligned or aligned(i) -- specifies alignment for the item.  Note that
           functions cannot have their alignment specified.  No argument
           means that we make it "maximally" aligned for the architecture.
  packed -- specifies the field, variable, or struct should be packed.

Note that the type-checker should take these into account when allowing
casts, but currently doesn't.

For variables and function declarations:
  shared -- used for DLL's where you want to share the same variable/fucntion
            across all instances of the dll.
  unused -- function/variable is unused (not enforced).
  weak   -- used for indirection at the linker level
  dllimport -- item is imported from a DLL (prepend _imp_ to the symbol)
  dllexport -- export the item for use as part of a DLL
  section("foo") -- place the item in the section "foo" in the object file

For function declarations:
  no_instrument_function -- used in conjunction with profiling
  constructor -- call the function at startup
  destructor -- call the function upon termination or exit
  no_check_memory_usage -- turn off memory leak debugging stuff