Cyclone now has malloc.  In C, malloc is a function from int to char *.  In Cyclone, malloc is a primitive that takes a type and returns a pointer to uninitialized heap space for a value of that type.  In order to look C-like and make porting easier, you write   malloc(sizeof(t)) for type t.  Notice that a cast of this to (t *) is redundant, but go ahead and do it to your heart's content.
You cannot allocate arrays of things this way.  Some day.  There is
parsing support for allocating enum variants this way, but the type
system is too weak to handle type-checking of this yet.

All of the infrastructure for this was already in place, this check-in
just handles the parsing, printing, etc. of the new expression form.

Here's an example file.  Uncomment the commented line and it will
fail to control-flow-check.

int b = 1;

int f() {
  int * x = malloc(sizeof(int));
  $(int,int*) @ y = malloc(sizeof($(int,int *)));
  (*y)[1] = x;
  if(b)
    *x = 17;
  else
    *(*y)[1] = 34;
  //  ++(*y)[0];
  if(b)
    return *(*y)[1];
  return *x;
}

I'd appreciate feedback on how much this helps in porting C code.