1. added support for bitfields.  One thing to note is that some bitfield members can have empty strings for names.  In this case, we consider the members to always be initialized (since you can't initialize them explicitly, nor get their values.)
2. re-organized the grammar to support "new e" instead of having
to write "new {e}".  This is in preparation for the change to
rewrite all of our "&Foo(e1,...,en)" and "&$(e1,...,en)" into
"new Foo(e1,...,en)" and "new $(e1,...,en)".  This change introduced
(or rather brought out a bug in) the code that tries to figure out
what the type of the pointer should be.  In particular, when you
write `new "foo"', there's an issue as to whether or not the "foo"
should be a char? or a char@{4}.  I've opted for the former (char?)
because that's what we normally need.  If you want the char@{4}
you have two options:  (1) write `new {"foo"}' instead or
(2) write `new (char@{4})"foo"' instead.

3. allow comprehensions as initializers for arrays.  So, you can
write the following:

  int x[10] = {for i < 10 : 0};

You can do this both at the top-level and within a function.  For
both cases, the upper bound must be a constant expression.  For
comprehensions within functions, we could allow the body of the
comprehension to be dynamically computed.  But I'm currently
restricting it to a constant to simplify the translation.