Added support for creating regions and allocating within them. The syntax is pretty straightforward:
region_t<`r> is the type of a region handle to region `r.

rnew(h) e allocates e within the region for which h is a handle.
rnew(h) {e1,...,en}  similarly but for an array
rnew(h) "foo"        similarly but for a string
rnew(h) {for i < e1 : e2} similarly but for a comprehension

rmalloc(h,sizeof(t)) similarly but doesn't initialize the value

To create a region, one writes:

  region<`r> h {
    stmt
  }

Within the braces, h has type region_t<`r> and `r must be a
fresh type variable and has kind RgnKind.  If you like, you
can leave off the <`r> and then the name of the region becomes
the same as the handle with a backquote (`) in front of it.
So, for instance if you write:

  region r {
   stmt
  }

then the type of r is region_t<`r>.  If an exception is raised
within stmt (or control is transferred via break, continue, or
goto outside the block) then the region should be freed.  I
haven't tested this yet though.

The runtime support for regions is minimal (see cyc_include.h and
runtime_cyc.c) and consists of:

// simple linked list of pages with variable length data size
struct _RegionPage {
  struct _RegionPage *next;
  char data[0];
};

// region handles
struct _RegionHandle {
  struct _RegionPage *curr; // pointer to first page in list of pages
  void *offset;             // pointer onto first page where to allocate
  void *last_plus_one;      // end of the first page (+ 1)
};


// allocates a new region with a default page of 128 bytes
struct _RegionHandle _new_region();

// allocates and returns pointer to object within region of size_t bytes
// rounded up to the minimal alignment constraints
void *_region_malloc(struct _RegionHandle *, size_t);

// frees all of the pages assocated with a region
void _free_region(struct _RegionHandle *);

Right now, we have a simple policy where each time we grow the region,
we add a page that is twice as big as the previous one (or the size of
the object plus some slack if that's bigger.)

The file tests/test_regions.cyc has a little test program that shows
how everything works.  I used it to verify that we actually do reclaim
the memory and run in constant space.