The most visible difference is that non-value-carrying xtunion variants are no longer heap-allocated.  So you say:   throw Dict::Absent and   catch { case Dict::Absent
That is why so many files are changed.  Notice that throwing
exceptions of this sort no longer causes heap-allocation.

The other visible difference at the Cyclone level is that xtunion
types can have type parameters.  You cannot extend the number of type
parameters later, for good reaons.

Internally, this is part of a greater change wherein tunion and
xtunion now share abstract syntax and a lot more of the code.  You can
look at the abstract-syntax for tunion and xtunion stuff to see the
details.

The translation to C changed substantively for xtunion: The unique
addresses are now 4-bytes of 0 followed by the name-string.  And
values of non-value-carrying variants are just these addresses.  Why
the 4-bytes of 0?  So that testing if a non-value-carrying variant is
a particular value-carrying-variant works -- you follow the address,
compare 0 against the value-carrying-variant's tag, and fail.

I have left interface.cyc broken -- an xtunion declaration will be
processed by extract_tuniondecl, but Matthieu suggests this is not
correct.

I'm still pretty unconvinced we get things like xtunion redeclarations
right, but we shouldn't be any buggier than before (famous last
words).

I have a little more clean-up to do that I'll check in soon (mostly in
the translation to C), but I shouldn't be in others' way anymore.