1. Added support for anonymous structs and unions.  Unlike C, we treat these structurally instead of gensym'ing a name and then treating them equivalent by name.  To use them, you have to use initializers.  That is, there's no constructor that generates an anonymous struct value.  It's not clear that my translation will actually go through GCC in all cases without giving warnings or errors since we are more expressive, but the test cases I've tried seem to work.
2. Fixed a bug w.r.t. regions and exception handling.  We were
popping off too many things from the handler stack when exiting
a region.  Also, break/continue within a region statement weren't
doing the right thing, so there was a potential space leak.  I've
re-organized how the handler stack works now to fix this.  Basically,
our handler stack has two different kinds of entries:  exception
handlers and region handles.  The latter are pushed when entering
a region block, and the former when entering a try block.  When
you pop out of a scope (either via continue, break, goto, or throw),
we pop off entries from the stack.  If they are region handle
entries, then we free the regions.  This way, no matter how control
is transferred outside of the region scope, we are sure to free
the region.  Also, this approach should make region blocks a little
cheaper as it avoids doing a setjmp and pushes the work off on
throwing an exception.