Fixed bug that Matthew found involving shadowed variables. Now when you have something like:
  int *i = new 3;
  int *i = i;

it won't type-check because this is treated the same as:

  int *i = new 3;
  int *i;
  i = i;

so i won't be initialized.  Note that for patterns, you can
write:

  int *i = new 3;
  let i = i;

the reason being that we use different variable names than
those occuring in the binders in the generated code.  However,
I had to turn off the optimization we had put in to make
let more efficient when we just had a variable because of
this issue.  So there's an unnecessary goto and label in
the generated code.  But that's better than a core dump :-)