Big change is a post-pass that eliminates uses of StmtExp.  This is ugly and not very robust, but the compiler can bootstrap with it turned on (flag -tovc, left off in the Makefiles).  The pass should always succeed for compiler-introduced StmtExp.  It may fail (mysteriously) if a (user-defined) StmtExp has jumps or returns out of the StmtExp.
The basic strategy is this:
 * Hoist declarations, renaming to avoid capture.  Do not hoist
 initializers, of course.
 * Where the declaration was, put an assignment.
 * Change statement sequences to expression sequences
 * Change if statements to conditional expressions

There are two ways this strategy can fail:
 * There are other statement forms (eg. for loops are the translation
 of comprehensions).
 * The initializer is not a legal expression (eg. the initializer for
 a tagged array is {e1, e2, e3}).

In both cases, I generate a static/inline function and replace the
statement/initializer with a call to the function.  The function's
arguments are the addresses of the free variables in the
statement/initializer.  The function's body is the
statement/initializer with the free variables being dereferenced.
Because the statement/initializer is now in a different function,
jumps and return outside of the statement/initializer make no sense.
Come to think of it, sizeof(arr) where arr is a local array probably
breaks too.

Other changes:

* Added some region support to Set.  We need this for Dict too.
* Some small changes to Toc to eliminate some unnecessary casts and
  clean up the code a bit.
* Changed printers to always surround SeqExp, i.e. e1,e2 in
  parentheses.  The problem is this is necessary in initializers and
  the printers don't know when they're in initializers
* Changed precedence of Increment in the printers.  I think this was
  just wrong.  It may explain some problems James had (we would print
  (*i)++ as *i++, which is equivalent to *(i++).