[Dan: hope I didn't undo any of your changes in toc.cyc but I don't have VC with me and couldn't verify that I didn't screw anything up. But my changes were pretty minimal so you should be able to see what's going on.]
Added support for varargs and made toc.cyc a bit more region-ified.

There are two versions of varargs: homogenous and heterogenous.  The
simplest is the homogenous and works as follows:

int sum(...`r int xs) {
  int s = 0;
  for (int i = 0; i < xs.size; i++) s++;
  return s;
}

The "...`r int" indicates that this function takes a variable number
of arguments of type int in the region `r.  From the callee's
standpoint, is no different from writing:

int sum(int ?`r xs) { ... }

but from the caller's standpoint, you can now write sum(),
sum(1,2,3), sum(1,2,3,4,5,6), etc.  The array is implicitly
allocated and initialized on the caller's stack and passed
to the callee.

This is pretty useful.  For instance, in the list module we now have:

  list_t<`a> list(...`r `a);

which is coded the same way as "from_array", but gives you a
convenient notation for allocating a list of known size such
as

  list(1,2,3,4)

instead of

  new List{1,new List{2,new List{3,new List{4,null}}}}

Of course, you can define a function that also takes other arguments
such as:

  void foo(t1 x, t2 y, t3 z, ... `r t argv)

but obviously the "..." has to occur at the end as with C's varargs.

Note that the region is required in the type.  This is mostly
due to parsing issues but also reminds the writer that the
array will be stack allocated, so I'm not too inclined to change
it.

To support a heterogeneous vararg, you have to use a combination of an
[x]tunion and the "inject" directive for the varargs.  For instance:

xtunion Printable {
  Double(d);
  Int(i);
  String(char?);
};

void print(...`r inject xtunion `r Printable xs) {
  for (; xs.size != 0; xs++)
    switch (*xs) {
    case &Double(d): printf("%d",d); break;
    case &Int(i): printf("%i",i); break;
    case &String(s): printf("\"%s\"",s); break;
    default: throw new InvalidArg("print");
    }
};

Here, xs has type "xtunion `r Printable ? `r" meaning that it's an
array of xtunion's and the xtunions are allocated in the same region
as the array.  The inject directive says that the caller should
(implicitly) inject arguments into the xtunion/tunion.  So, for
instance, you could write:

  print();
  print("foo");
  print(3);
  print(3.1415,"foo",3)

etc.  The goal is to allow us enough flexibility to write printf,
scanf, etc. in the language.

The rules for injection are a little hacky.  Basically, at the point
of use, we try to find a constructor of the appropriate type.  We
first try to find an exact match with no coercions.  Nonetheless, this
can constrain types because unification is used.  If no constructor
matches through unification, then we then try to match again using
only silent coercions.  This allows you to do things like treat "foo"
as a char? instead of a char@{4}.  Other than that, we don't do any
other coercions.  (I tried supporting others, but it's very
confusing.  This was the best compromise I could find...)  The first
constructor that matches according to the scheme above is picked
and there's no check to see that this match is unique.

So you should write as many unique cases as possible in the
union/xtunion to avoid amiguities.  For instance, calling the above
print with a char might be bad news because it can be silently coerced
to a double, even though you might think it'll be coerced to an int.
The obvious solution is to add a Char(char) constructor.

xtunion's add a bit of a twist because constructors come into and go
out of scope.  The basic rule is that we inject according to the
constructors in scope at the call site (not the definition site of the
function.)  This gives us a crude facility for building extensible
methods, though you have to code up your own dictionaries and do a lot
of try/catch tricks to get it to work out right.  I coded one of these
up to check it out.  Interestingly, it required first-class
polymorphism to get the effects right...

In the future, I expect to make inject orthogonal from the varargs
stuff, but it wasn't worth it for now.  I also need to add some way to
convert a vararg function to a normal function.  And I need to add
support for statically checking things like printf against a
descriptor string literal where possible.  And it might be useful
to leave the region off and have that mean you want the array to be
heap-allocated by the client instead of stack-allocated.  Also, we
really need not require that the array and [x]tunion values be in the
same region (though they always are, if we allow the function to be
coerced to a non-vararg type, then there's no need for the constraint.)

Finally, we should rewrite all of the printf, vsnprintf, etc. that the
C standard library provides using this new facility.  Or rather, we
should port something like what Dan found for snprintf.

There's a way to do the latter so that we can add new printable
or parsable types -- exercise left to the reader :-)