Finally!  We now have basic socket support, and a stripped-down BSD finger is in tests/.  The C version is there, too, so you can diff the two and see how much change was required to port it.
One Cyclone bug seems to have been revealed, and it's pointed out in
finger.cyc: with a variable "host" of type "char ?", the test "if
(!host)" generates a gcc error "aggregate value used where an integer
was expected".  Using "if (host == null)" works, but both forms work
if "host" has type "struct hostenv *", for example.

The Cyclone implementation of errno now calls a C function to retrieve
a pointer to the real errno rather than trying to steal it from the
Windows DLL, which wasn't working under Linux anyway.

One other problem appeared earlier today, but I can't reproduce it
now: the Cyclone call strerror(errno) was returning the string
"Success" even when the connect() call in finger had just failed.  I
figured some code in the wrapper somewhere was calling a library
function and inadvertently changing errno, but I can't figure it out.

As promised, now Cyclone does not escape into the native C header
files, and "#include <...>" is the preferred form for inclusion of
standard headers.  I have changed "..." to <...> where appropriate in
lib/, src/, and some of tests/, but I have not done this in tools/.