We now allow you to cast a pointer into region r1 to a pointer into region r2 as long as r1 outlives r2.  The heap region lives as long as any region.  Region parameters live longer than any locally-declared region, etc.  This is implemented by keeping in the environment a partial order on region variables. I think this extra degree of flexibility is really needed to type-check more C code.  For instance, you can do something like this now:
int x = 0;
if (foo) {
  int y = 3;
  int *_ z = friday ? ((int *_)&x) : &y;
  *z++;
}

Right now, the explicit cast is required to avoid a warning, but
we could change that if people like.

The next step is to add some way to specify the partial order on
function types, so that you can propagate the informataion across
calls.  We could also augment existentials so that if you have
a value like:  Exists r<:r2.int*r, then you could open up the
existential, cast the int*r to an int*r2, and then if r2 is in
your capability, access the pointer.  (Note that r will never
be in your capability because you opened the thing up, thereby
alpha-converting it to some "fresh" region variable that can't
be in your capability.)