If your program doesn't use the garbage collector, why link it in?  Use -nogc instead.
With gc and -s, hello world is 53760 bytes.
With -nogc and -s, it is 8704 bytes.
In C with -s it is 3072 bytes.

Every program will still do 1 malloc for argv and
1 malloc for each string in argv, regardless of -nogc.

To see how many mallocs your program did with -nogc,
declare

extern "C" nogc_report();

and call it just before exiting.