* Changed cyc_include.h like this: #ifndef _throw extern int _throw(void * e); #endif
This allows a wrapper to do something like:

#define _throw(X) exit(0)

and then not need to link in the Cyclone library.  (If you don't
ignore X, you'll get that Null_Exception isn't defined at link-time.)

I'm open to suggestions for "less dangerous" ways to allow this sort
of thing.  A certain build environment for device drivers doesn't seem
longjmp friendly. (If you're curious, it's not exit-friendly either,
but there's a function for causing an immediate blue-screen.)

If this change is deemed "totally evil", then I can just make my own
cyc_include.h.

* Bug fix in printing calling conventions for MSVC.  There are still
  some strange places where attributes seem to be getting silently dropped.

* Small Toc changes.