Two smallish changes:
1. cases within a switch that have an empty statement are treated
as if there was a fallthru;.  This means that you can write things
like:

case 'a': case 'b': case 'c': <do-something>

which I found to be common in the printf and scanf code I ported
from gcc.  I think it shows up in many other places as well.

2. the functions terr, impos, etc. within the compiler were changed
to use varargs.  I then went through and got rid of all of the
terr(loc,aprintf(...)) calls and replaced them with terr(loc,...).
That should make the code a bit smaller in various places.