Another up optimization that makes a big difference (about 20%) on Trevor's generated files: <X></X> can be eliminated for any tag X.
Originally I thought the way to test for this was a Text such that the
pos held '\0', so up_opt now takes the buf_t.  Turns out the parser
instead hands me an empty Seq, so I'm not actually using the buf_t.
I've commented out the call to get_char to test for '\0', but if it's
possible to occur, then I can re-enable it.