Added a new generate algorithm.  Usage is now
  generate [-alg] [len]

where alg is 0 or 1; 0 is the original and 1 is the new.

smlng doesn't compress the output much, on the other hand
the output might not permit much compression.

More seriously, if I run

  ./generate -1 100 | tee XX | ./smlng 180 | tee XXX | ./validate XX

I generally find the validator says there is an error in
the output of smlng.  Typically it involves an output character
that should have the original color or size (indicated by * in
the output of validate), but the corresponding character of XXX
does not have the original color or size.