OK here is code to detect the stack limit.  I hope the stack grows DOWN on their machine :-).
If you make, then run ./stack, it will try to predict the stack limit
and verify it by overflowing.  Hopefully it prints a small number of
lines of addresses as it gets near.  Give it a try on a machine near
you.

The prediction code is in a function at the beginning of the file
stack.cyc.  I'll have to let someone else hook this in as my brain is
becoming fried.  I will try to test the script...