Fixed the following bugs:
* using Foo; using Foo; caused ambiguous definition error.

* now allow more modifiers on descriptor strings in printf. Still
need to do scanf.

* the last case in a switch no longer requires a break, and cases
that implicitly fall through from one case to the next, where the
next case introduces no pattern variables gets a warning instead
of an error.

* no longer allow x -= y when x and y are both pointers, since
the result of x-y is an int.

* now deal with "int x[] = {for i < 3 : ...};"  That is, you
no longer have to put in the size for both the type and in the
comprehension.

Since this was all done under Win98, there may be some ^M's
that are spurious.  Sorry about that...