Fixed bug in the configure script.  In addition to defining the the architecture's standard #define's in config/arch/ARCH, it also includes in that file the architecture's name ARCH, but with the - and .'s removed.  This is useful for setting up platform-specific cyclone header files (i.e. for signals, etc.).
Also fixed some things with the interface checker.  In particular, I
made some things extern that were not (like declarations to xtunion exn),
and relaxed the interface checking rules so that structs need not be
extern, to help out with C compatibility.  There is still more work
to do here, though.