Added Makefile target update_all_archs:
For each supported architecture ARCH, the src and lib directories are
cleaned out, new .c files are compiled for ARCH, and these are installed
in bin/genfiles/ARCH.  To support this, I made the configure script
figure out all of the supported architectures and stored them in the
Makefile.inc file.  Therefore, you should remove your Makefile.inc
and reconfigure when you update to this.

I didn't remove the redundant .h files in lib/---I'll leave that to
someone that's more motivated than me. :)

Also, the changes made to interface.cyc didn't make it into the last
check in (that support no extern on the front of structs).