I added support for storing architectures as patches.  As an example, I included a port to Sun.
(Everyone needs to redo ./configure ...)

The build procedure is the same as before: make everything under
your own architecture for as long as you like, but NEVER
DO A MAKE UPDATE DIRECTLY.  Instead, once you are ready to commit,
do a make update_all_archs; I have modified its behavior to update
all supported architectures, including your own.

The magic is in the update_all_archs target.  This
routine cycles through every architecture, and extracts it from
its patch, if a patch for it exists.  Then, for every architecture,
it does a make update, thus propogating changes for all architectures.
Finally, for those architectures that were originally patches,
it regenerates the patch from the now-updated reference
architecture, which is i686-unknown-linux (the reference architecture
name is stored in the variable PATCH_ARCH, defined in Makefile.inc).
Finally, it removes the directories for the architectures that
have patches, just leaving the patch behind.

In terms of savings, the .c files for Sun were roughly 4 MB,
while the patch is only 15K.

Notes:

In truth, doing a make update is fine if you are not developing
on Linux.  If you are on Linux, doing a make update will change
the files upon which the patches for other architectures
depend, so this is a no no.

If folks start to feel like supporting lots of architectures in
this way is a hassle, we can easily avoid it by truncating the
ALL_ARCHS variable in Makefile.inc to be just linux and cygwin.
Then we can build patches and so forth for other archs only
at release time.  However, this means we have to ultimately
build on some other architecture each time we do a release to
regenerate the .c files.