Various tweaks to support an OpenBSD port:
1) If bash is not installed, it is safe to replace it in the
configure and config/cyc_install scripts with /bin/ksh instead. I
noted this in the INSTALL document.
2) The gc did not build due to a snafu in architecture determination.
I made *very small* change to gc/include/private/gcconfig.h to fix
this---Trevor, I hope this doesn't screw up licensing.
3) BSD ar did not support the -s flag, so I removed it from all of
the Makefiles and did a ranlib immediately following the ar
instead.
4) Added the files need to support the BSD port.

In addition, I tweaked the top-level makefile so that if your
architecture is shipped as a patch, it will leave the unpacked
directory in genfiles, rather than removing it after repatchifying
it.

I added target update_devel_archs so that when we do check-ins to
the compiler, we only have to run update_devel_archs to fix up
Linux and Cygwin; update_all_archs only need be used before a release.