Wow, lots of changed files.  Most of these just have been changed to add a new absyn case, as I now describe.
I've added the beginnings of support for basic type-directed functions, like
marshallers.  There is a new syntactic primitve called __gen that
takes a type argument and (when implemented) "returns" a value of
type tunion Typestruct (currently defined in tests/typerep.h; this file
should be moved eventually).  In typerep.h, a macro maps gen() to __gen()
so that users are "encouraged" to include the typerep.h file, which
is needed for the translation to work.  Using Typestruct, we can define
a marshalling function like

void marshal(tunion Typestruct,`a, FILE @);

which would be called as follows, using gen():

marshal(gen(t),v,f)

where v is some value having type t.  We can also define an unmarshalling
function, a structural equality function, etc. using this facility.
(Actually, this won't work because gen() is restricted to appear in the
initialization expression of top-level functions to simplify the translation).

Right now, the code that is to implement the translation from gen(t)
to a value is in tcgenrep.cyc.  The entry-point function returns a list
of declarations and an expression; the typechecker calls this function
and then splices in the declarations in the top-level list, and replaces
the gen() expression with the one returned.  At the moment tcgenrep.cyc
just has the scaffolding in place to do the translation, but it is
not performed; this will be done soon.