NOTE: change to Makefile.inc.in; please re-run configure.
Some small changes:

1) The XML library is now built during normal installation and
will be installed if the user does make install.  I do not store
the generated .c files in genfiles, but I do store the header
files in include/.  If you don't want to build the xml library
by default, set the flag NO_XML_LIB in Makefile.inc.

2) Added some helper routines to xml.cyc for processing XML
documents.

3) Added routine "remove" to the queue library that will let
you pull an given element out of the middle of a queue. I
only committed genfiles for cygwin.  Anyone else who wants it
has to build it for their platform and do the update.