A rewrite of the control-flow checker.  It no longer uses "abstract expressions", so re-run configure -- a file has been removed.
The code accepts all our code and rejects "really obvious" uses of
uninitialized variables.  I haven't tested it exhaustively by any
means, so let me know about problems.  It should be more
understandable than the last version, but it still needs a ton of
documentation.

The new analysis computes "is zero" or "is not zero" or "unknown" for
all locals.  This isn't being used yet, but as soon as I decorate the
abstract syntax with the answers, the translation to C can use it to
eliminate null-checks.  This optimization is especially powerful
because an earlier pointer-dereference means (under appropriate
assumptions) that a later one can't cause a NullException.
We can also use this analysis to avoid spurious warnings in the
type-checker, but that's a bit of a nuissance because the type-checker
runs first.

I assume the big slowdown in compiling absyn.cyc is because of the
type-rep stuff.  But let me know if it was me. :-)