Added some more comments to absyn.h.  We ought to re-organize this in a top-down fashion.
Added a "guide to porting from C" section to the manual.  There's
some duplication with the FAQ, but I think it's good to have a
list of things that you know you'll have to do.  There's also
some details in here about how to build wrappers and write
interfaces to existing C code.